consolidated all epics & ophyd setup into 1 file
