+++ src/instrument/core/run_engine.py
@@ -17,11 +17,13 @@ from ..configs.loaders import iconfig
 logger = logging.getLogger(__name__)
 logger.info(__file__)
 
-from .best_effort import bec
-from .catalog import cat
-from .epics_setup import connect_scan_id_pv
+from ..utils.controls_setup import connect_scan_id_pv
+from ..utils.controls_setup import set_control_layer
+from ..utils.controls_setup import set_timeouts
 from ..utils.metadata import MD_PATH
 from ..utils.metadata import re_metadata
+from .best_effort import bec
+from .catalog import cat
 
 re_config = iconfig.get("RUN_ENGINE", {})
 
@@ -41,7 +43,10 @@ RE.subscribe(cat.v1.insert)
 RE.subscribe(bec)
 RE.preprocessors.append(sd)
 
-connect_scan_id_pv(RE)  # if configured
+set_control_layer()
+set_timeouts()  # MUST happen before ANY EpicsSignalBase (or subclass) is created.
+
+connect_scan_id_pv(RE)  # if configured #TODO: What does this mean?
 
 if re_config.get("USE_PROGRESS_BAR", True):
     # Add a progress bar.
